pharmacy_inventory/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.19
django-taggit==6.1.0
djangorestframework==3.15.2
greenlet==1.1.2
gunicorn==23.0.0
idna==2.8
mysqlclient==2.1.0
numpy==1.15.0
packaging==24.2
pep8==1.7.1
pillow==10.4.0
psutil==5.9.0
pycodestyle==2.8.0
PyGObject==3.36.0
python-apt==2.0.0+ubuntu0.20.4.7
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
SQLAlchemy==1.4.32
sqlparse==0.5.3
ssh-import-id==5.10
typing-extensions==4.12.2
unattended-upgrades==0.1
urllib3==1.25.8
